revert naming of Touch to ZProbe, add additional user commands, user files
add ProbeXYZ User command
Add toolbutton to each User command and probe commands to open a file
selector dialog and let the user pick a file to be used for commands.

Change 'sendMacro' to check if the passed string is a file url and
open the file instead of sending commands directly.

rename 'Z Probe' to Z Touch to match the UI (changes translations too)

add ugly icon for XYZ probing