abs/rel work/machine toggles in rpncalc
add two toggle buttons to rpn calc, absolute/relative and work/machine

this simplifies the keypad and makes the axis buttons a little simpler
to follow.

The functions of the '->Axis', 'Axis->', and 'GO Axis' buttons depend
on the state of the toggles.  Hopefully this is mostly intuitive, with
the following exception.

When "Machine" is selected, the ->Axis buttons are disabled because
there isn't really a way to set the machine axis - this is what homing
is for.